# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RFA knowledge server — /ask
  summary: desk(대응 에이전트) ↔ 지식 서버 계약
  description: 'head → broker → task agent → censor 를 수행하는 `ask()` 하나를 노출한다. `audience` 가 검열 프로파일을 정한다(public→public,
    company/self→internal). 동기 응답(서버 타임아웃 180초 → refusal no_knowledge/timeout). 같은 request_id 는 캐시 응답.
    `feedback[].reason` 은 사람 입력으로 신뢰되어 검열 규칙(learned.yaml)으로 되먹임된다. 인증: `Authorization: Bearer <RFA_ASK_TOKEN>`
    (.env.dev).'
  version: 1.0.0
paths:
...
  /tasks:
    get:
      tags:
      - frontend
      summary: List Tasks
      description: '좌측 태스크 목록: ask.yaml 카탈로그 + 만든 태스크(팀). 생성 중(applying)도 포함, 실패는 제외.'
      operationId: listTasks
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskView'
                title: Response Listtasks
    post:
      tags:
      - frontend
      summary: Create Task
      description: '태스크 추가(태스크 1 = 에이전트 1). owner 전용. 검증 실패는 스트림 전에 JSON 오류(403/409/422/503), 통과하면 SSE
        로 요구사항 분석 → 팀 설계(패턴 선택) → spawning 을 보낸다. 연결이 끊겨도 생성은 계속되며 `GET /tasks/{id}` 로 확인한다. 봉투 `{type,
        runId, seq, ts, data}`; type: `task.start` {taskId, name, agentName}, `task.stage` {stage: analyze|design|spawn,
        status: running|done|error, ms, detail} — analyze.detail {capabilities: [{id, label}], source},
        design.detail {roles, supervisor, members: [{agentId, role}], sandbox}, spawn.detail {agentsApply,
        seeded}, `task.log` {stage, text}, `task.done` {task: TaskView, agent: AgentView, ms}, `task.error`
        {stage, code: no_role_for_requirement|invalid_team|apply_failed|task_create_failed, message, detail}'
      operationId: createTask
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: SSE stream of SseEnvelope frames
          content:
            application/json:
              schema: {}
            text/event-stream:
              schema:
                description: 'Every SSE event of every stream: ``{type, runId, seq, ts, data}``. ``type``
                  names may change

                  with the front-end; the envelope does not.'
                properties:
                  type:
                    description: event name, e.g. run.started | stage | delta | guard.final | done | error
                    title: Type
                    type: string
                  runId:
                    title: Runid
                    type: string
                  seq:
                    description: 0-based, strictly increasing within a run
                    minimum: 0
                    title: Seq
                    type: integer
                  ts:
                    description: unix seconds
                    title: Ts
                    type: number
                  data:
                    additionalProperties: true
                    title: Data
                    type: object
                required:
                - type
                - runId
                - seq
                - ts
                title: SseEnvelope
                type: object
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /tasks/{task_id}:
    get:
      tags:
      - frontend
      summary: Get Task
      description: 태스크 하나. 생성 스트림이 끊겼을 때 applying → ready|failed 를 확인한다(failed 도 돌려준다).
      operationId: getTask
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Color:
      properties:
        bg:
          type: string
          title: Bg
          description: '#rrggbb background'
        fg:
          type: string
          title: Fg
          description: '#rrggbb foreground'
      type: object
      required:
      - bg
      - fg
      title: Color
...
    ErrorOut:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ErrorOut
      description: 'Front-end error body: ``code`` for the client, ``message`` ready to show (Korean).'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskCreateRequest:
      properties:
        name:
          type: string
          maxLength: 30
          title: Name
          description: 태스크명 (비면 422 task_name_required)
          default: ''
        agentName:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Agentname
          description: 비면 「{태스크명} 대응 에이전트」
        description:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Description
          description: 비면 「{태스크명} 태스크를 맡은 에이전트입니다.」
        tags:
          items:
            type: string
          type: array
          maxItems: 32
          title: Tags
          description: '≤8 kept; # stripped, trimmed, de-duplicated. A tag in a question routes it here'
      type: object
      title: TaskCreateRequest
      description: '``POST /tasks`` (「태스크 추가」). No 사내/사외 grade: it belongs to each request, not the task.'
...
    TaskView:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
          description: task name (left-nav title)
        agentId:
          type: string
          title: Agentid
          description: its chat partner (`GET /agents`)
        agentName:
          type: string
          title: Agentname
        desk:
          type: string
          title: Desk
          description: left-nav subtitle
        icon:
          type: string
          enum:
          - star
          - github
          - research
          - slack
          - mail
          - shield
          - generic
          title: Icon
          default: generic
        color:
          $ref: '#/components/schemas/Color'
        initials:
          type: string
          title: Initials
          default: ''
        itemCount:
          type: integer
          title: Itemcount
          default: 0
        status:
          type: string
          enum:
          - ready
          - applying
          - failed
          title: Status
          default: ready
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        source:
          type: string
          enum:
          - catalog
          - team
          title: Source
          default: catalog
        worker:
          anyOf:
          - type: string
          - type: 'null'
          title: Worker
          description: backend agent that answers (task agent or team supervisor)
        sandbox:
          anyOf:
          - type: string
          - type: 'null'
          title: Sandbox
        securityLabel:
          type: string
          title: Securitylabel
          description: sandbox network privilege (admin), not the 사내/사외 grade
          default: ''
      type: object
      required:
      - id
      - name
      - agentId
      - agentName
      - desk
      - color
      title: TaskView
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError